Fix storage-report-test
