+++ apps/studio/tests/pages/projects/reports/storage-report.test.js
@@ -4,43 +4,20 @@ import { waitFor, screen } from '@testing-library/react'
 import { StorageReport } from 'pages/project/[ref]/reports/storage'
 import userEvent from '@testing-library/user-event'
 
-// [Joshen] Am temporarily commenting out the breaking tests due to:
-// "TypeError: _fetch.get.mockReset is not a function" error from Jest
-// just so we get our jest unit/UI tests up and running first
-// Need to figure out how to mock the "get" method from lib/common/fetch properly
+// [Joshen] Mock data for ApiReport is in __mocks__/hooks/useStorageReport
+// I don't think this is an ideal set up as the mock data is not clear in this file itself
+// But this is the method that worked for me after hours of wrangling with jest.spyOn and jest.mock
+// which for some reason none of them worked when I was trying to mock the data within the file itself
+// I'd be keen to see how we can do this better if anyone is more familiar to jest 🙏
 
-beforeEach(() => {
-  // reset mocks between tests
-  // get.mockReset()
-  // get.mockImplementation(async (_url) => [{ result: [] }])
-})
-
-test(`static elements`, async () => {
+test(`Render static elements`, async () => {
   render(<StorageReport />)
   await screen.findByText('Request Caching')
   await screen.findByText(/Last 24 hours/)
 })
 
-test('refresh button', async () => {
+test('Render top cache misses', async () => {
   render(<StorageReport />)
-  // await waitFor(() => expect(get).toBeCalled())
-  // get.mockReset()
-  // userEvent.click(await screen.findByText(/Refresh/))
-  // await waitFor(() => expect(get).toBeCalled())
-})
-
-test('append - top cache misses', async () => {
-  // get.mockImplementation(async (url) => {
-  //   if (decodeURIComponent(url).includes('misses')) {
-  //     return {
-  //       result: [{ path: 'mypath', search: 'some-query', count: 22 }],
-  //     }
-  //   }
-  //   return { result: [{ timestamp: new Date().toISOString(), miss_count: 123, hit_count: 123 }] }
-  // })
-  // render(<StorageReport />)
-  // await waitFor(() => expect(get).toBeCalled())
-  // await screen.findAllByText(/mypath/)
-  // await screen.findAllByText(/some\-query/)
-  // await screen.findAllByText(/22/)
+  await screen.findAllByText('/storage/v1/object/public/videos/marketing/tabTableEditor.mp4')
+  await screen.findAllByText('2')
 })
